build: Package core functions in ct-node for now. No build process. Revisit post release
